+++ lib/chat_web/live/main_live/layout/cargo_room.ex
@@ -4,11 +4,10 @@ defmodule ChatWeb.MainLive.Layout.CargoRoom do
   use ChatWeb, :component
 
   alias Chat.Sync.CargoRoom
-  alias ChatWeb.MainLive.Layout.Timer
 
   def bar(assigns) do
     ~H"""
-    <%= if @media_settings.functionality == :cargo && @cargo_room && @cargo_room.pub_key == @room.pub_key do %>
+    <%= if @cargo_room && @cargo_room.pub_key == @room.pub_key do %>
       <div class="w-full px-8 py-4 border-b border-white/10 backdrop-blur-md bg-grayscale/40 z-10 flex flex-row items-center justify-between text-md text-white">
         <div class="flex flex-row items-center text-base">
           Cargo sync activated <.remove_button status={@cargo_room.status} />
@@ -26,8 +25,6 @@ defmodule ChatWeb.MainLive.Layout.CargoRoom do
       <div>
         Insert empty USB drive
       </div>
-
-      <Timer.timer timer={@cargo_room.timer} />
     </div>
     """
   end