+++ lib/chat_web/live/main_live/page/room.ex
@@ -4,10 +4,12 @@ defmodule ChatWeb.MainLive.Page.Room do
   use ChatWeb, :component
 
   import ChatWeb.MainLive.Page.Shared
+  import ChatWeb.LiveHelpers.LiveModal
+  import ChatWeb.LiveHelpers, only: [open_content: 0]
   import Phoenix.Component, only: [assign: 3]
 
   import Phoenix.LiveView,
-    only: [consume_uploaded_entry: 3, push_event: 3, send_update: 2]
+    only: [consume_uploaded_entry: 3, push_event: 3, send_update: 2, push_patch: 2]
 
   require Logger
 
@@ -23,26 +25,30 @@ defmodule ChatWeb.MainLive.Page.Room do
   alias Chat.Messages
   alias Chat.RoomInviteIndex
   alias Chat.Rooms
-  alias Chat.Rooms.{Registry, Room, RoomRequest}
+  alias Chat.Rooms.{Registry, Room, RoomMessageLinks, RoomRequest}
   alias Chat.Sync.{CargoRoom, UsbDriveDumpRoom}
   alias Chat.Upload.UploadMetadata
   alias Chat.User
+  alias Chat.Utils
   alias Chat.Utils.StorageId
 
-  alias ChatWeb.MainLive.Layout
+  alias ChatWeb.MainLive.Page
 
   alias Phoenix.PubSub
 
   @per_page 15
+  @usb_drive_dump_progress_topic "chat::usb_drive_dump_progress"
 
   def init(socket), do: socket |> assign(:room, nil)
 
   def init(%{assigns: %{room_map: rooms}} = socket, room_key) when is_binary(room_key) do
-    room = Rooms.get(room_key)
-    room_identity = rooms |> Map.fetch!(room_key)
-
-    socket
-    |> init({room_identity, room})
+    with %Room{} = room <- Rooms.get(room_key),
+         %Identity{} = room_identity <- Map.get(rooms, room_key) do
+      init(socket, {room_identity, room})
+    else
+      _ ->
+        socket
+    end
   end
 
   def init(
@@ -61,13 +67,63 @@ defmodule ChatWeb.MainLive.Page.Room do
     |> assign(:edit_content, nil)
     |> assign(:room, room)
     |> assign(:room_identity, room_identity)
+    |> assign(:is_room_linked?, RoomMessageLinks.has_room_linked_messages?(room_identity))
     |> assign(:last_load_timestamp, nil)
     |> assign(:has_more_messages, true)
     |> assign(:message_update_mode, :replace)
+    |> assign(:usb_drive_dump_room, UsbDriveDumpRoom.get())
     |> assign_messages()
     |> assign_requests()
     |> maybe_enable_cargo()
     |> maybe_enable_usb_drive_dump()
+    |> push_event("chat:scroll-down", %{})
+  end
+
+  def init_with_linked_message(socket, hash) do
+    with {ciphered_identity, _, msg_index, msg_id} <- RoomMessageLinks.get(hash),
+         identity <-
+           Rooms.decipher_identity(ciphered_identity, hash |> Base.decode16!(case: :lower)),
+         room <- identity |> Identity.pub_key() |> Rooms.get() do
+      socket
+      |> store_new(identity)
+      |> init({identity, room})
+      |> load_messages_to({msg_index, msg_id})
+      |> send_js(open_content())
+      |> push_patch(to: "/")
+    else
+      _ ->
+        socket
+        |> assign(:lobby_mode, :rooms)
+        |> init()
+        |> push_event("chat:toggle", %{to: "#chatRoomBar", class: "hidden"})
+        |> push_event("chat:toggle", %{to: "#navbarLeft", class: "navbar"})
+        |> push_event("chat:toggle", %{to: "#navbarLeft", class: "hidden"})
+        |> push_event("chat:toggle", %{to: "#navbarTop", class: "navbarTop"})
+        |> push_event("chat:toggle", %{to: "#navbarTop", class: "hidden"})
+        |> push_event("chat:toggle", %{to: "#navbarBottom", class: "navbarBottom"})
+        |> push_event("chat:toggle", %{to: "#navbarBottom", class: "hidden"})
+        |> push_event("chat:toggle", %{to: "#contentContainer", class: "hidden"})
+    end
+  end
+
+  def store_key_copy(%{assigns: %{me: me, room_map: room_map}} = socket, room_identity) do
+    unless Map.has_key?(room_map, Identity.pub_key(room_identity)) do
+      my_notes = Dialogs.find_or_open(me)
+
+      room_identity
+      |> Messages.RoomInvite.new()
+      |> Dialogs.add_new_message(me, my_notes)
+      |> RoomInviteIndex.add(my_notes, me)
+    end
+
+    socket
+  end
+
+  def store_new(socket, new_room_identity) do
+    socket
+    |> store_key_copy(new_room_identity)
+    |> Page.Login.store_new_room(new_room_identity)
+    |> Page.Lobby.refresh_room_list()
   end
 
   def load_more_messages(%{assigns: %{page: page}} = socket) do
@@ -331,6 +387,7 @@ defmodule ChatWeb.MainLive.Page.Room do
 
   def close(%{assigns: %{room: room}} = socket) do
     PubSub.unsubscribe(Chat.PubSub, room.pub_key |> room_topic())
+    PubSub.unsubscribe(Chat.PubSub, @usb_drive_dump_progress_topic)
 
     socket
     |> assign(:room, nil)
@@ -350,6 +407,49 @@ defmodule ChatWeb.MainLive.Page.Room do
     msg_id
     |> Rooms.read_message(room_identity)
     |> maybe_redirect_to_file(socket)
+  end
+
+  def link_message(
+        %{assigns: %{room: room, room_identity: room_identity}} = socket,
+        {index, id},
+        render_fun
+      ) do
+    :ok = RoomMessageLinks.create(room, room_identity, {index, id})
+
+    socket
+    |> assign(:is_room_linked?, true)
+    |> push_event("chat:change", %{
+      to: "#room-message-#{id} .link-status",
+      content: render_to_html_string(%{linked: true, msg_id: id, msg_index: index}, render_fun)
+    })
+    |> forget_current_messages()
+  end
+
+  def unlink_messages_modal(socket, component) do
+    socket |> open_modal(component)
+  end
+
+  def unlink_messages(%{assigns: %{room_identity: room_identity}} = socket, render_fun) do
+    RoomMessageLinks.cancel_room_links(room_identity)
+    |> Enum.reduce(socket, fn {_, _, index, id}, socket ->
+      content = render_to_html_string(%{linked: false, msg_id: id, msg_index: index}, render_fun)
+
+      socket
+      |> push_event("chat:change", %{to: "#room-message-#{id} .link-status", content: content})
+    end)
+    |> assign(:is_room_linked?, false)
+    |> close_modal()
+  end
+
+  def share_message_link_modal(%{assigns: %{}} = socket, msg_id, component) do
+    message_url =
+      [ChatWeb.Endpoint.url(), "room", RoomMessageLinks.link_hash(msg_id)] |> Path.join()
+
+    socket
+    |> open_modal(component, %{
+      url: message_url,
+      encoded_qr_code: Utils.qr_base64_from_url(message_url)
+    })
   end
 
   defp maybe_redirect_to_file(%{type: type, content: json}, socket)
@@ -387,18 +487,18 @@ defmodule ChatWeb.MainLive.Page.Room do
   end
 
   def open_image_gallery(socket, msg_id) do
-    send_update(Layout.ImageGallery, id: "imageGallery", action: :open, incoming_msg_id: msg_id)
+    send_update(Page.ImageGallery, id: "imageGallery", action: :open, incoming_msg_id: msg_id)
     socket
   end
 
   def image_gallery_preload_next(socket) do
-    send_update(Layout.ImageGallery, id: "imageGallery", action: :preload_next)
+    send_update(Page.ImageGallery, id: "imageGallery", action: :preload_next)
 
     socket
   end
 
   def image_gallery_preload_prev(socket) do
-    send_update(Layout.ImageGallery, id: "imageGallery", action: :preload_prev)
+    send_update(Page.ImageGallery, id: "imageGallery", action: :preload_prev)
 
     socket
   end
@@ -442,7 +542,34 @@ defmodule ChatWeb.MainLive.Page.Room do
     |> assign(:room_requests, request_list)
   end
 
-  defp assign_requests(socket), do: socket
+  defp assign_requests(socket), do: socket |> assign(:room_requests, [])
+
+  defp load_messages_to(%{assigns: %{has_more_messages: false}} = socket, {_, msg_id}) do
+    socket
+    |> push_event("chat:scroll", %{to: "#message-block-#{msg_id}"})
+    |> push_event("chat:toggle", %{to: "#message-block-#{msg_id}", class: "bg-black/10"})
+  end
+
+  defp load_messages_to(
+         %{
+           assigns: %{
+             room: room,
+             room_identity: identity,
+             last_load_timestamp: index,
+             messages: messages
+           }
+         } = socket,
+         {msg_index, msg_id}
+       ) do
+    prev_messages = Rooms.read_to(room, identity, {index - 1, 0}, {msg_index, msg_id})
+    messages = prev_messages ++ messages
+
+    socket
+    |> assign(:messages, messages)
+    |> assign(:last_load_timestamp, set_messages_timestamp(messages))
+    |> push_event("chat:scroll", %{to: "#message-block-#{msg_id}"})
+    |> push_event("chat:toggle", %{to: "#message-block-#{msg_id}", class: "bg-black/10"})
+  end
 
   defp broadcast_message_updated(msg_id, pub_key, me, time) do
     {:updated_message, msg_id}
@@ -521,6 +648,8 @@ defmodule ChatWeb.MainLive.Page.Room do
   end
 
   def maybe_enable_usb_drive_dump(%{assigns: %{room: room}} = socket) when not is_nil(room) do
+    PubSub.unsubscribe(Chat.PubSub, @usb_drive_dump_progress_topic)
+
     usb_drive_dump =
       cond do
         match?(%CargoRoom{status: :syncing}, socket.assigns[:cargo_room]) ->
@@ -531,6 +660,7 @@ defmodule ChatWeb.MainLive.Page.Room do
           when pub_key == room.pub_key or status == :dumping,
           socket.assigns[:usb_drive_dump_room]
         ) ->
+          PubSub.subscribe(Chat.PubSub, @usb_drive_dump_progress_topic)
           :disabled
 
         true ->